Share Rails 8.1 adapter behavior
